Do not tag images with rocker-... or rocker__detection to allow for automatic pruning
Docker images need to be cleaned up from time to time to not fill up the
file system with obsolete images. This is typically done by running
`docker image prune` or `docker system prune`. This command does not
remove tagged images by default.

In fact, there is no need for rocker to assign tags to the intermediate
images. The image IDs work just fine.

+++ src/rocker/os_detector.py
@@ -17,11 +17,11 @@ import pexpect
 from ast import literal_eval 
 from io import BytesIO as StringIO
 
-from .core import get_docker_client
+from .core import docker_build
 
 
-DETECTOR_DOCKERFILE="""
-FROM python:3
+DETECTION_TEMPLATE="""
+FROM python:3 as detector
 
 RUN mkdir -p /tmp/distrovenv
 RUN python3 -m venv /tmp/distrovenv
@@ -32,10 +32,6 @@ RUN echo 'import distro; print(distro.linux_distribution())' > /tmp/distrovenv/d
 RUN . /tmp/distrovenv/bin/activate && pyinstaller --onefile /tmp/distrovenv/detect_os.py
 
 RUN . /tmp/distrovenv/bin/activate && staticx /dist/detect_os /dist/detect_os_static
-"""
-
-DETECTION_TEMPLATE="""
-FROM rocker__detector as detector
 
 FROM %(image_name)s
 
@@ -45,37 +41,13 @@ CMD [ "" ]
 """
 
 
-def build_detector_image(verbose=False):
-    client = get_docker_client()
-    """Build the image to use to detect the OS"""
-    dockerfile_tag = 'rocker__detector'
-    iof = StringIO(DETECTOR_DOCKERFILE.encode())
-    im = client.build(fileobj = iof, tag=dockerfile_tag)
-    log = []
-    for l in im:
-        log.append(l.decode())
-    success = False
-    # Check for success. sometimes it's the last sometimes it tags last.
-    for l in log[-2:-1]:
-        if "Successfully built " in l:
-            success = True
-    if not success:
-        print("Failed to build image %s:\n>>>>\n%s\n>>>>>" % (dockerfile_tag, '\n'.join(log)))
-    elif verbose:
-        print("Successfully built image %s:\n>>>>\n%s\n>>>>>" % (dockerfile_tag, '\n'.join(log)))
-    return success
-
 def detect_os(image_name):
-    client = get_docker_client()
-    dockerfile_tag = 'rocker__detection_%s' % image_name
     iof = StringIO((DETECTION_TEMPLATE % locals()).encode())
-    im = client.build(fileobj = iof, tag=dockerfile_tag)
-    for l in im:
-        pass
-        #print(l)
-    
+    image_id = docker_build(fileobj = iof)
+    if not image_id:
+        return None
 
-    cmd="docker run -it --rm %s" % dockerfile_tag
+    cmd="docker run -it --rm %s" % image_id
     p = pexpect.spawn(cmd)
     output = p.read()
     p.terminate()
